cloud_topics/gc: extract dependency interfaces to namespace scope
Move nested interface classes (object_storage, epoch_source,
node_info, safety_monitor), the collection_error and state enums
out of level_zero_gc into a separate level_zero_gc_types.h header
under namespace cloud_topics::l0::gc, matching the convention used
by the rest of the level_zero/ code.

Preparation for templatizing level_zero_gc with a Clock parameter:
these types don't depend on Clock and must be accessible without
knowing the Clock type.
